+++ src/layout/section/heroSection/index.unit.tsx
@@ -4,28 +4,26 @@ import { render, screen } from '@testing-library/react'
 
 import { HeroSection, HeroSectionProps } from './heroSection'
 
-function createProps(props: Partial<HeroSectionProps>): HeroSectionProps {
+function createProps(props: Partial<HeroSectionProps> = {}): HeroSectionProps {
   return {
     heroText: 'hero text',
     heroDescription: 'hero description',
-    heroImageUrlLarge: 'http://heroImageUrl',
+    heroImageUrl: 'http://heroImageUrl',
     ...props,
   }
 }
 
 describe('HeroSection', () => {
-  it('should render basic hero section', () => {
-    const props = createProps({
-      heroImageUrl: 'http://heroImageUrl',
-      buttonText: 'button text',
-      buttonHref: 'http://buttonhref',
-    })
-
+  it('should render the title', () => {
+    const props = createProps()
     render(<HeroSection {...props} />)
-
     expect(screen.getByText(props.heroText)).toBeInTheDocument()
+  })
+
+  it('should render the desciption', () => {
+    const props = createProps()
+    render(<HeroSection {...props} />)
     expect(screen.getByText(props.heroDescription)).toBeInTheDocument()
-    expect(screen.getByText(props.buttonText)).toHaveAttribute('href', props.buttonHref)
   })
 
   it('should render a search form', () => {